+++ src/com/intellij/idea/plugin/hybris/tools/remote/console/impl/HybrisPolyglotQueryConsole.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -37,7 +37,6 @@ import javax.swing.Icon
 import javax.swing.JPanel
 import javax.swing.JSpinner
 import javax.swing.SpinnerNumberModel
-import javax.swing.border.EmptyBorder
 
 class HybrisPolyglotQueryConsole(project: Project) : HybrisConsole(project, HybrisConstants.CONSOLE_TITLE_POLYGLOT_QUERY, PolyglotQueryLanguage.instance) {
 
@@ -67,13 +66,13 @@ class HybrisPolyglotQueryConsole(project: Project) : HybrisConsole(project, Hybr
     }
 
     private fun initCommitElements() {
-        commitLabel.border = EmptyBorder(labelInsets)
+        commitLabel.border = JBUI.Borders.empty(labelInsets)
         panel.add(commitLabel)
         panel.add(commitCheckbox)
     }
 
     private fun initMaxRowsElements() {
-        maxRowsLabel.border = EmptyBorder(labelInsets)
+        maxRowsLabel.border = JBUI.Borders.empty(labelInsets)
         panel.add(maxRowsLabel)
         panel.add(maxRowsSpinner)
     }